Allow exp claim to have string value #3202
This commit changes how the HTTP response is processed. It ensures that error handling is done when a string is passed as exp value in the IntrospectResponse.

This change checks the type of the exp value when the Introspect response object is created. The previous version accepted only a string. But this commit accepts a string and tries to parse it to an int and assign it